// ADW app with AEV/assets/source-code/adw-3.1.1-aev/libs/content-ee/governance/src/lib/record/models/declare-record.model.ts
/*
 * Copyright © 2005 - 2021 Alfresco Software, Ltd. All rights reserved.
 *
 * License rights for this program may be obtained from Alfresco Software, Ltd.
 * pursuant to a written agreement and any use of this program without such an
 * agreement is prohibited.
 */

import { Node } from '@alfresco/js-api';
export interface DeclareRecord {
    entry: Node;
    status: 'failed' | 'passed';
}